third_party/WebKit/LayoutTests/presentation/presentation-reconnect.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/gc.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/presentation-service-mock.js"></script>
<button>click me</button>
<script>

async_test(t => {
  presentationServiceMock.then(mockService => {
    var button = document.querySelector('button');
    var connection = null;
    var request = new PresentationRequest('https://example.com');

    waitForClick(_ => {
      request.start().then(conn => {
        connection = conn;
        assert_not_equals(connection, null);

        request.reconnect(connection.id).then(
          t.step_func_done(conn => {
            assert_true(connection === conn);
        }));
      });
    }, button);
  });
}, "Test that Presentation.reconnect() resolves with existing presentation connection.");

</script>
</body>
</html>
